정건희: <style> 마이페이지 전적 디자인 수정 반영, 업적 기본 css 적용

+++ src/components/myPage/myPageSignOut/MyPageSignOut.jsx
@@ -11,13 +11,16 @@ function MyPageSignOut() {
       <button
         className="myPageSignOutBtn fontSemiBold"
         onClick={() => {
-          instance.delete('lier/removal').then((res) => {
-            alert(res.data.data);
-          });
-          sessionStorage.clear();
-          new Cookies().remove('access_token');
-          new Cookies().remove('refresh_token');
-          navigate('/');
+          const result = window.confirm('정말로 탈퇴하시겠습니까?');
+          if (result) {
+            instance.delete('lier/removal').then((res) => {
+              alert(res.data.data);
+            });
+            sessionStorage.clear();
+            new Cookies().remove('access_token');
+            new Cookies().remove('refresh_token');
+            navigate('/');
+          }
         }}
       >
         회원탈퇴
